requirements.txt
fastapi==0.121.2
faiss-cpu==1.9.0.post1
httpx==0.28.1
matplotlib==3.10.7
mcp==1.21.2
networkx==3.5
numpy==2.3.5
pandas==2.3.3
plotly==6.3.1
python-dotenv==1.2.1
PyYAML==6.0.3
Requests==2.32.5
rich==14.2.0
seaborn==0.13.2
sentence-transformers==3.3.1
sse_starlette==3.0.2
uvicorn==0.38.0
jinja2==3.1.4
psutil==5.9.8
